# docs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Secret:
      description: "Represents a torero secret.\n\nSecrets are sensitive values stored\
        \ securely and used by torero\nservices and decorators.\n\nAttributes:\n \
        \   name: Unique identifier for the secret\n    description: Human-readable\
        \ explanation of the secret's purpose\n    type: Category of secret (e.g.,\
        \ 'password', 'api-key', 'token')\n    created_at: Timestamp when the secret\
        \ was created\n    metadata: Optional additional metadata about the secret"
      example:
        created_at: '2023-01-01T00:00:00Z'
        description: Database password for the application
        metadata:
          owner: admin
          provider: vault
        name: db-password
        type: password
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Timestamp when the secret was created
          title: Created At
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable explanation of the secret's purpose
          title: Description
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Additional metadata about the secret
          title: Metadata
        name:
          description: Unique identifier for the secret
          title: Name
          type: string
        type:
          description: Category of secret (e.g., 'password', 'api-key', 'token')
          title: Type
          type: string
      required:
      - name
      - type
      title: Secret
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  contact:
    email: opensource@itential.com
    name: torero Development Team
    url: https://torero.dev/contact
  description: "\n        RESTful API for interacting with torero services.\n    \
    \    \n        This API provides endpoints for discovering and managing torero\
    \ services,\n        allowing you to list services, filter by type or tag, and\
    \ get metadata\n        about available service types and tags.\n        \n  \
    \      Model Context Protocol (MCP) Integration:\n        This API follows OpenAPI\
    \ standards and provides comprehensive type\n        information for seamless\
    \ integration with MCP-enabled applications.\n        "
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: torero API
  version: 0.1.0
  x-mcp-compatible: true
  x-mcp-description: This API is optimized for Model Context Protocol integration.
  x-mcp-version: '1.0'
openapi: 3.1.0
paths:
...
  /v1/secrets/:
    get:
      description: "Get all registered torero secrets with optional filtering.\n \
        \   \n    This endpoint returns a list of all secrets registered with torero.\n\
        \    You can filter the results by secret type to narrow down the list.\n\
        \    \n    Examples:\n    - List all secrets: GET /v1/secrets/\n    - List\
        \ all password secrets: GET /v1/secrets/?type=password\n    - List all API\
        \ key secrets: GET /v1/secrets/?type=api-key"
      operationId: list_secrets_v1_secrets__get
      parameters:
      - description: Filter by secret type, e.g. 'password', 'api-key', 'token'
        in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by secret type, e.g. 'password', 'api-key', 'token'
          title: Type
      - description: Number of items to skip
        in: query
        name: skip
        required: false
        schema:
          default: 0
          description: Number of items to skip
          minimum: 0
          title: Skip
          type: integer
      - description: Maximum number of items to return
        in: query
        name: limit
        required: false
        schema:
          default: 100
          description: Maximum number of items to return
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Secret'
                title: Response List Secrets V1 Secrets  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List secrets
      tags:
      - secrets
...
  /v1/secrets/{name}:
    get:
      description: "Get detailed information about a specific secret by name.\n  \
        \  \n    This endpoint retrieves metadata about a single secret identified\
        \ by its name.\n    The name is case-sensitive and must match exactly the\
        \ name of a registered secret.\n    \n    Note: For security reasons, this\
        \ endpoint only returns metadata about the secret,\n    not the actual secret\
        \ value.\n    \n    If no secret is found with the specified name, a 404 error\
        \ is returned."
      operationId: get_secret_v1_secrets__name__get
      parameters:
      - description: Name of the secret to retrieve
        in: path
        name: name
        required: true
        schema:
          description: Name of the secret to retrieve
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secret'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get secret details
      tags:
      - secrets
...